output running version on console message
